6S980-hw2/requirements.txt
einops
jaxtyping
beartype
black
matplotlib
ruff
scipy
trimesh
pytest
tqdm
hydra-core
